refactor: List -> Set 으로 변경
